🗺️ refactor: Update component file paths
