+++ src/app/study/[id]/edit/page.tsx
