+++ src/app/study/[id]/focus/Header.tsx
