+++ src/app/study/[id]/focus/Point.tsx
