+++ src/app/study/[id]/focus/Timer.tsx
