+++ src/app/study/[id]/focus/page.tsx
