+++ src/app/study/[id]/page.tsx
