+++ src/app/study/layout.tsx
