+++ src/components/study/RecentStudyBoard.tsx
