+++ src/components/study/SearchInput.tsx
