+++ src/components/study/StudyCard.tsx
