+++ src/components/study/StudyExplorerBoard.tsx
